adding changes to backend

+++ server/Api/Models/Domain/Product.cs
@@ -11,7 +11,7 @@ namespace Webshop.Models.Domain
         public int Id { get; set;}
         public string ProductClass { get; set; }
         public string ProductName { get; set; }
-        public int UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         //public int AantalOpOrder { get; set; } 
       //  public int? Amount { get; set; }
         public int Availability { get; set; }
@@ -26,7 +26,7 @@ namespace Webshop.Models.Domain
         public bool InStock { get; set; }
         #region ctors
        
-        public Product(string productclass, string productname,int unitPrice,string description
+        public Product(string productclass, string productname,double unitPrice,string description
            // , int? amount=null
             ):this()
         {
